Lock migrate_plus to 5.2
New releases on `migrate_plus` break the patching we are calling for here.  There is a new 5.3 version and a new 6.0 version, but we haven't tested either with this module or the rest of the stack yet. This is to get tests to pass on the profile until we can fully test the new versions with the rest of our software.